Restructure config and app logic
