+++ api/v2/app/interfaces/job/manticore/core-image-settings.js
