+++ api/v2/app/interfaces/job/manticore/hmi-image-settings.js
