+++ api/v2/app/interfaces/job/manticore/index.js
